Extract common functionality into abstract class.
